CR pass: fix file-naming inconsistency, drop indirection, merge test-only file splits
- Rename provider_cmd.go/rule_cmd.go to provider_command.go/rule_command.go
  to match the _command.go convention already used by agent/cc/profile.
- Inline runAdd's []string-arg indirection into ProviderAddCmdKong.Run
  now that the Kong struct already carries the four typed fields — the
  wrapper only existed from the pre-Kong CLI dispatch era.
- Same for RuleImportCmdKong: runImport now takes a plain file string
  instead of a single-element []string built just to pass through.
- RuleExportCmdKong.Run already fetches the rule by UUID; runExport used
  to re-fetch it by request-model+scenario. Pass the already-fetched
  *typ.Rule through instead of doing the lookup twice.
- Merge token_view_all_test.go + token_kind_test.go into token_test.go
  (same pattern as the earlier provider/rule test-file merge — two files
  only existed because of incremental history, not a real boundary).
- Move TestStandaloneBotSettingPreservesClaudeProfileSelection out of
  agent_routing_test.go into remote_test.go: it tests remote.go's
  standaloneBotSetting, not agent code — the old file name and its
  stale doc comment were left over from a prior refactor.
- gofmt token.go (struct tag alignment drifted after an earlier edit).
