+++ internal/command/rule_command_test.go
